Revert "[llcpp] StringView + fidl::tracking_ptr"
This reverts commit dc5212d78e639094cca738f647d68cf0eb2dcccc.

Reason for revert: CI build failure

Original change's description:
> [llcpp] StringView + fidl::tracking_ptr
> 
> - StringView is implemented via VectorView
> - fidl::unowned_str, fidl::heap_copy_str, fidl::copy_str to create
> strings
> - Literals can directly construct StringView
> 

TBR=bprosnitz@google.com,ianloic@google.com,apang@google.com,pascallouis@google.com

No-Presubmit: true
No-Tree-Checks: true
No-Try: true

+++ zircon/system/ulib/fidl/include/lib/fidl/llcpp/allocator.h
@@ -92,7 +92,7 @@ class Allocator {
     void* data;
     bool requires_delete;
   };
-  virtual allocation_result allocate(size_t obj_size, size_t count, destructor dtor) = 0;
+  virtual allocation_result allocate(size_t obj_size, uint32_t count, destructor dtor) = 0;
 
  private:
   template <typename T>
